requirements.txt
Django==4.2.0
djangorestframework==3.14.0

mssql-django==1.6
pillow==10.0.0
python-decouple==3.8
django-filter==23.2
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.5
numpy>=1.24.0
django-cors-headers>=4.3.0
django-extensions
opencv-python
tensorflow
django-cryptography
mediapipe